Anjay-freertos-client 23.02
Features
- Reworked socket implementation to use X-Cube-Cellular offloaded sockets
  (Com sockets) instead of LwIP (NOTE: currently supporting only UDP)
- Added Firmware Update Object (/5) to STM32L496G-DISCOVERY and B-L462E-CELL1 projects

Improvements
- Revamped configuration of Anjay and its dependencies
- Updated Anjay to version 3.3.0

+++ Middlewares/ST/STM32_Cellular/Interface/Com/Src/com_sockets_ip_modem.c
@@ -578,9 +578,9 @@ static bool com_translate_ip_address(const com_sockaddr_t *p_addr, int32_t addrl
         {
           com_ip_addr_t com_ip_addr;
           com_ip_addr.addr = ((const com_sockaddr_in_t *)p_addr)->sin_addr.s_addr;
-          (void)sprintf((CSIP_CHAR_t *)p_socket_addr->ip_value, "%hhu.%hhu.%hhu.%hhu",
-                        COM_IP4_ADDR1(&com_ip_addr), COM_IP4_ADDR2(&com_ip_addr),
-                        COM_IP4_ADDR3(&com_ip_addr), COM_IP4_ADDR4(&com_ip_addr));
+          (void)sprintf((CSIP_CHAR_t *)p_socket_addr->ip_value, "%u.%u.%u.%u",
+                        (unsigned int)COM_IP4_ADDR1(&com_ip_addr), (unsigned int)COM_IP4_ADDR2(&com_ip_addr),
+                        (unsigned int)COM_IP4_ADDR3(&com_ip_addr), (unsigned int)COM_IP4_ADDR4(&com_ip_addr));
         }
         p_socket_addr->port = COM_NTOHS(((const com_sockaddr_in_t *)p_addr)->sin_port);
         result = true;
